+++ src/main/java/listener/MyChangeListListener.java
@@ -3,8 +3,10 @@ package listener;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vcs.changes.ChangeListListener;
 import service.ViewService;
+import system.Defs;
 
 public class MyChangeListListener implements ChangeListListener {
+    private static final com.intellij.openapi.diagnostic.Logger LOG = Defs.getLogger(MyChangeListListener.class);
 
     private final ViewService viewService;
     public MyChangeListListener(Project project) {
@@ -12,6 +14,7 @@ public class MyChangeListListener implements ChangeListListener {
     }
 
     public void changeListUpdateDone() {
+        LOG.debug("changeListUpdateDone() called - triggering update");
         // TODO: collectChanges: VcsTree is updated
         viewService.incrementUpdate();
         viewService.collectChanges(true);